AI8/requirements.txt
# Core dependencies for HRNet Facial Keypoints Detection
torch>=1.9.0
torchvision>=0.10.0
opencv-python>=4.5.0
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.4.0
scipy>=1.7.0
scikit-image>=0.18.0
Pillow>=8.3.0
tqdm>=4.62.0

# Optional dependencies for better performance
tensorboard>=2.6.0
albumentations>=1.1.0

# For COCO format handling
pycocotools>=2.0.2

# For visualization
seaborn>=0.11.0
